Move normalizeFilenameID
